+++ Unigram/Unigram/Views/InstantPage.xaml.cs
@@ -1597,7 +1597,7 @@ namespace Unigram.Views
         {
         }
 
-        public void OpenUser(int userId)
+        public void OpenUser(long userId)
         {
         }
 
@@ -1609,7 +1609,7 @@ namespace Unigram.Views
         {
         }
 
-        public void OpenViaBot(int viaBotUserId)
+        public void OpenViaBot(long viaBotUserId)
         {
         }
